third_party/WebKit/LayoutTests/external/wpt/xhr/overridemimetype-edge-cases.window-expected.txt

This is a testharness.js-based test.
PASS overrideMimeType() is not reset by open(), basic
PASS overrideMimeType() is not reset by open()
PASS If charset is not overridden by overrideMimeType() the original continues to be used
FAIL Charset can be overridden by overrideMimeType() with a bogus charset assert_equals: expected "\ufffd\ufffd" but got "รรฐ"
Harness: the test ran to completion.
